pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "boneio-backend"
version = "0.1.0"
description = "Backend package for BoneIO"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic<2",
    "asyncio-mqtt",
    "pyyaml",
    "minimalmodbus",
]

[tool.setuptools.packages.find]
namespaces = true

